requirements.txt
flask
flask-login
flask-sqlalchemy
flask-bcrypt
flask-restful
flask-restx
flask-mail
flask-wtf
flask-admin
email-validator
pymysql
cryptography
requests
flask-cors
bs4
flask-marshmallow
libsass
flask-migrate
python-slugify
